add EducationalForm component, update PersonalForm
still needs an 'add' function to add entries
to educational and work forms.
labels need to be adjusted and a date range put in
for years. not very dry because its mostly a copy
of PersonalForm.

update structure of the custom form components,
now input fields can easily be placed over the <p>
that replace them when editActive=false.