add EducationalForm component, update PersonalForm
Still needs an 'add' function to add entries
to educational (and work) forms.
Labels need to be adjusted and a date range put in
for years. Not very dry because it's mostly a copy
of PersonalForm.

Update structure of the custom form components:
Input fields can be placed better over the <p>
elements that replace them when editActive=false.