update handleSubmit, fix unique key error
Add the code for updating data categories that
don't have an 'entries' property. Will change so
each category's data is stored in an entry, even
if there is only one.

setParentValues doesn't work, the submitted Form's
childData is updated (the values are rendered) but
parentValues is the same before and after submit.