switch to using proper keys, fix editActive state
Just as planned, using the array index for setting
keys of child elements in a list lead to unwanted
behaviour when changing the index of the elements.
Now however, every Form is wrapped in a div which
has a proper unique key.
This key also ensures that the editActive state
stays on an element when its index is changed.